+++ app/src/main/java/com/novoideal/tabuademares/controller/SeaConditionController.java
@@ -47,7 +47,7 @@ public class SeaConditionController {
             return;
         }
 
-        SeaCondition[] today = new SeaCondition[3];
+//        SeaCondition[] today = new SeaCondition[3];
 
         for (SeaCondition condition : result) {
             if(Days.daysBetween(new DateTime(condition.getDate()), new DateTime(city.getDate())).getDays() != 0){
@@ -55,23 +55,23 @@ public class SeaConditionController {
             }
             switch (condition.getPeriod()){
                 case "manha" :
-                    today[0] = condition;
-//                    updateCondiction(condition, R.id.a_m, R.id.s_m, R.id.w_m);
+//                    today[0] = condition;
+                    updateCondiction(condition, R.id.a_m, R.id.s_m, R.id.w_m);
                     break;
                 case "tarde" :
-                    today[1] = condition;
-//                    updateCondiction(condition, R.id.a_t, R.id.s_t, R.id.w_t);
+//                    today[1] = condition;
+                    updateCondiction(condition, R.id.a_t, R.id.s_t, R.id.w_t);
                     break;
                 case "noite" :
-                    today[2] = condition;
-//                    updateCondiction(condition, R.id.a_n, R.id.s_n, R.id.w_n);
+//                    today[2] = condition;
+                    updateCondiction(condition, R.id.a_n, R.id.s_n, R.id.w_n);
                     break;
             }
         }
 
-        if (today[0] != null) {
-            createGridView(today);
-        }
+//        if (today[0] != null) {
+//            createGridView(today);
+//        }
 
     }
