+++ app/src/main/java/com/novoideal/tabuademares/model/LocationParam.java
@@ -161,7 +161,7 @@ public class LocationParam {
 
     @Override
     public String toString() {
-        return name;// + " - " + new DateTime(getDate()).toString("dd/MM/yyyy");
+        return "Previsão para: " + name;// + " - " + new DateTime(getDate()).toString("dd/MM/yyyy");
     }
 
     public String getTodayStr() {
