+++ app/src/main/res/drawable/rounded_corner.xml
@@ -2,10 +2,8 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
 
     <stroke
-        android:width="1dp"
-        android:color="@color/colorPrimary" />
-
-    <solid android:color="@color/extremeTide" />
+        android:width="0.1dp"
+        android:color="@color/colorPrimaryDark" />
 
     <gradient
         android:startColor="@color/colorPrimary"
