+++ app/src/main/res/drawable/tab_layout.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape  xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient
+        android:endColor="#CCCCCC"
+        android:angle="90"
+        android:startColor="#ffffff" />
+
+</shape>
