+++ app/src/main/res/drawable/table_border.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:height="2dp"
+        android:left="3dp"
+        android:right="3dp"
+        android:top="0dp">
+        <shape android:shape="rectangle">
+            <gradient
+                android:angle="360"
+                android:startColor="@color/extremeTide"
+                android:endColor="#006386"
+                android:type="linear" />
+        </shape>
+    </item>
+
+</layer-list>
