+++ app/src/main/res/layout/activity_main.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.design.widget.CoordinatorLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_content"
@@ -9,12 +10,12 @@
     android:fitsSystemWindows="true"
     tools:context="com.novoideal.tabuademares.MainActivity">
 
-    <!--<android.support.design.widget.AppBarLayout-->
-        <!--android:id="@+id/appbar"-->
-        <!--android:layout_width="match_parent"-->
-        <!--android:layout_height="wrap_content"-->
-        <!--android:paddingTop="@dimen/appbar_padding_top"-->
-        <!--android:theme="@style/AppTheme.AppBarOverlay">-->
+    <android.support.design.widget.AppBarLayout
+        android:id="@+id/appbar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingTop="@dimen/appbar_padding_top"
+        android:theme="@style/AppTheme.AppBarOverlay">
 
         <!--<android.support.v7.widget.Toolbar-->
             <!--android:id="@+id/toolbar"-->
@@ -25,8 +26,19 @@
             <!--app:popupTheme="@style/AppTheme.PopupOverlay">-->
 
         <!--</android.support.v7.widget.Toolbar>-->
+    <android.support.design.widget.TabLayout
+        android:id="@+id/tabs"
+        android:background="@drawable/tab_layout"
+        android:layout_width="match_parent"
+        app:tabTextColor="#d3d3d3"
+        app:tabSelectedTextColor="#0baeff"
+        android:layout_height="40dp"
+        app:tabGravity="fill"
+        app:tabMode="fixed" />
 
-    <!--</android.support.design.widget.AppBarLayout>-->
+    </android.support.design.widget.AppBarLayout>
+
+
 
     <android.support.v4.view.ViewPager
         android:id="@+id/container"
