+++ app/src/main/res/layout/spinner_item.xml
@@ -4,4 +4,5 @@
     android:layout_height="wrap_content"
     android:textSize="20dp"
     android:text="Cabo Frio"
+    android:padding="5dp"
     android:textColor="#191970" />
