+++ app/src/main/res/layout/table_swell.xml
@@ -1,0 +1,156 @@
+<TableLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/tableLayout1"
+    android:layout_marginBottom="50sp"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:stretchColumns="0,1,2,3"
+    >
+
+    <TableRow
+        android:id="@+id/tableRow1"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:padding="5dip">
+
+        <TextView
+            android:text="Período"
+            android:textAlignment="center"
+            android:textColor="#368BC1"
+            android:textSize="12dp"
+            android:textStyle="bold" />
+
+        <TextView
+            android:id="@+id/morning"
+            android:textAlignment="center"
+            android:textColor="#368BC1"
+            android:text="@string/agitation"
+            android:textSize="12dp"
+            android:textStyle="bold" />
+
+        <TextView
+            android:id="@+id/afternoon"
+            android:textAlignment="center"
+            android:textColor="#368BC1"
+            android:text="@string/lbl_swell"
+            android:textSize="12dp"
+            android:textStyle="bold"/>
+
+        <TextView
+            android:id="@+id/night"
+            android:textAlignment="center"
+            android:textColor="#368BC1"
+            android:text="@string/lbl_wind"
+            android:textSize="12dp"
+            android:textStyle="bold" />
+    </TableRow>
+
+    <TableRow
+        android:id="@+id/tableRow2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:background="@drawable/table_border"
+        android:padding="5dip">
+
+        <TextView
+            android:id="@+id/head_1"
+            android:textAlignment="center"
+            android:text="@string/morning_12h"
+            android:textSize="12dp"
+            android:textStyle="bold" />
+
+        <TextView
+            android:id="@+id/a_m"
+            android:textAlignment="center"
+            android:textStyle="bold"
+            android:text="@string/loading"
+            android:textSize="12dp" />
+
+        <TextView
+            android:id="@+id/a_t"
+            android:textStyle="bold"
+            android:textAlignment="center"
+            android:text="@string/loading"
+            android:textSize="12dp" />
+
+        <TextView
+            android:id="@+id/a_n"
+            android:textStyle="bold"
+            android:textAlignment="center"
+            android:text="@string/loading"
+            android:textSize="12dp" />
+    </TableRow>
+
+    <TableRow
+        android:id="@+id/tableRow3"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:background="@drawable/table_border"
+        android:padding="5dip">
+
+        <TextView
+            android:id="@+id/head_2"
+            android:textAlignment="center"
+            android:textSize="12dp"
+            android:text="@string/afternoon_18h"
+            android:textStyle="bold" />
+
+        <TextView
+            android:id="@+id/s_m"
+            android:textStyle="bold"
+            android:textAlignment="center"
+            android:text="@string/loading"
+            android:textSize="12dp" />
+
+        <TextView
+            android:id="@+id/s_t"
+            android:textStyle="bold"
+            android:textAlignment="center"
+            android:text="@string/loading"
+            android:textSize="12dp" />
+
+        <TextView
+            android:id="@+id/s_n"
+            android:textStyle="bold"
+            android:textAlignment="center"
+            android:text="@string/loading"
+            android:textSize="12dp" />
+    </TableRow>
+
+    <TableRow
+        android:id="@+id/tableRow4"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:background="@drawable/table_border"
+        android:padding="5dip">
+
+        <TextView
+            android:id="@+id/head_3"
+            android:textAlignment="center"
+            android:textSize="12dp"
+            android:text="@string/night_21h"
+            android:textStyle="bold" />
+
+        <TextView
+            android:id="@+id/w_m"
+            android:textStyle="bold"
+            android:textAlignment="center"
+            android:text="@string/loading"
+            android:textSize="12dp" />
+
+        <TextView
+            android:id="@+id/w_t"
+            android:textStyle="bold"
+            android:textAlignment="center"
+            android:text="@string/loading"
+            android:textSize="12dp" />
+
+        <TextView
+            android:id="@+id/w_n"
+            android:textStyle="bold"
+            android:textAlignment="center"
+            android:text="@string/loading"
+            android:textSize="12dp" />
+    </TableRow>
+
+
+</TableLayout>
