+++ app/src/main/res/layout/weather_condiction.xml
@@ -1,53 +1,61 @@
-<?xml version="1.0" encoding="utf-8"?>
-<GridLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<TableLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/tableLayout1"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:columnCount="3"
-    android:orientation="horizontal"
-    android:paddingTop="5dp"
-    android:rowCount="2"
-    android:rowOrderPreserved="true">
+    android:stretchColumns="0,1,2">
 
-    <ImageView
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:src="@drawable/sun" />
-
-    <ImageView
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:src="@drawable/moon" />
-
-    <ImageView
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:src="@drawable/wind" />
-
-    <TextView
-        android:id="@+id/weather_main"
+    <TableRow
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:maxWidth="100dp"
-        android:singleLine="false"
-        android:text="Ensolarado com pancadas de chuva de manhã" />
+        android:padding="5dip">
 
-    <TextView
-        android:id="@+id/moon_phase"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:maxWidth="100dp"
-        android:text="Gibosa minguante"
-        android:textAlignment="center" />
+        <ImageView
+            android:layout_gravity="center"
+            android:gravity="center"
+            android:src="@drawable/sun" />
 
-    <TextView
-        android:id="@+id/weather_wind"
+        <ImageView
+            android:layout_gravity="center"
+            android:gravity="center"
+            android:src="@drawable/moon" />
+
+        <ImageView
+            android:layout_gravity="center"
+            android:gravity="center"
+            android:src="@drawable/wind" />
+    </TableRow>
+
+    <TableRow
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:maxLines="3"
-        android:maxWidth="100dp"
-        android:text="Nordeste, 10 nós"
-        android:textAlignment="center" />
-</GridLayout>
+        android:padding="0dip">
+
+        <TextView
+            android:id="@+id/weather_main"
+            android:layout_gravity="center"
+            android:gravity="center"
+            android:maxWidth="100dp"
+            android:singleLine="false"
+            android:text="Ensolarado com pancadas de chuva de manhã" />
+
+        <TextView
+            android:id="@+id/moon_phase"
+            android:layout_gravity="center"
+            android:gravity="center"
+            android:maxWidth="100dp"
+            android:singleLine="false"
+            android:text="Gibosa minguante"
+            android:textAlignment="center" />
+
+        <TextView
+            android:id="@+id/weather_wind"
+            android:layout_gravity="center"
+            android:gravity="center"
+            android:maxLines="3"
+            android:maxWidth="100dp"
+            android:singleLine="false"
+            android:text="Nordeste, 10km/h"
+            android:textAlignment="center" />
+
+    </TableRow>
+</TableLayout>
